move all error definitions to errors package
