+++ sdk/errors/plugin.go
@@ -1,11 +1,9 @@
-package sdk
+package errors
 
 import (
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 )
-
-// FIXME: move to the 'errors' package
 
 // UnsupportedCommandError is an error that can be used to designate that
 // a given device does not support an operation, e.g. write. Write is required
@@ -26,12 +24,12 @@ func (e *EnumerationNotSupported) Error() string {
 	return "This plugin does not support device auto-enumeration."
 }
 
-// invalidArgumentErr creates a gRPC InvalidArgument error with the given description.
-func invalidArgumentErr(format string, a ...interface{}) error {
+// InvalidArgumentErr creates a gRPC InvalidArgument error with the given description.
+func InvalidArgumentErr(format string, a ...interface{}) error {
 	return status.Errorf(codes.InvalidArgument, format, a...)
 }
 
-// notFoundErr creates a gRPC NotFound error with the given description.
-func notFoundErr(format string, a ...interface{}) error {
+// NotFoundErr creates a gRPC NotFound error with the given description.
+func NotFoundErr(format string, a ...interface{}) error {
 	return status.Errorf(codes.NotFound, format, a...)
 }
